Initial commit
We explain the purpose of this repository, the history behind it, where
it's going & how can others help. We also captured all contributors to
date.

We also captured the Varnish + Fly.io config as it existed for:
- Let's build a CDN - Part 1 - https://www.youtube.com/watch?v=8bDgWvyglno
- Kaizen 15 - NOT a pipe dream - https://changelog.com/friends/50

There is a lot more context here: https://github.com/thechangelog/changelog.com/pull/518

The most interesting part is the `run` script. To run it, you will need to:
1. Have a Fly.io account
2. Have a back-end app deployed 💡 https://fly.io/speedrun/
3. Change the name of the backend app (i.e. `changelog-2024-01-12`)
4. Launch the app in this repository

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>