Add dynamic backend, closest to the origin, cache-status (#1)
This is the second iteration which:
- uses dynamic backends
- places the initial instance closest to the origin
- improves on the x-cache header via cache-status

We also made the VARNISH_SIZE smaller so that it fits the default
instance size. We are leaving room for "sidecar" processes (log drains,
purge worker, etc.) and also for testing what happens when the memory
fills up.

We did this together with James A Rosen & Matt Johnson on August 2,
2024. Will add the recording when it's public. It's a follow-up to:
- https://github.com/thechangelog/changelog.com/pull/518

We also talked about this in a YouTube Live Stream with Peter Mbanugo:
https://www.youtube.com/live/ghE59eB465I

Signed-off-by: Gerhard Lazu <gerhard@changelog.com>